Move some functionality from ClientWindow to ExtendedClient
